allow tdust_warm to be free parameter; testing
Changed calc_500um_excess.py and two_temp_model_bayes.py to allow
tdust_warm to be a free parameter. Prior for tdust_warm is uniform
between 50 and 100 K. calc_500um_excee.py changed to make the fitting
of all of the sources to be a function rather than script. Image files
are from testing the changes.